Init React, Mostly Finished Express

+++ todos/functions/APIs/todos.js
@@ -2,6 +2,7 @@ const { db } = require("../util/admin");
 
 exports.getAllTodos = (request, response) => {
   db.collection("todos")
+    .where("username", "==", request.user.username)
     .orderBy("createdAt", "desc")
     .get()
     .then((data) => {
@@ -33,6 +34,7 @@ exports.postOneTodo = (request, response) => {
 
   const newTodoItem = {
     title: request.body.title,
+    username: request.user.username,
     body: request.body.body,
     createdAt: new Date().toISOString(),
   };
@@ -56,6 +58,9 @@ exports.deleteTodo = (request, response) => {
     .then((doc) => {
       if (!doc.exists) {
         return response.status(404).json({ error: "Todo not found" });
+      }
+      if (doc.data().username !== request.user.username) {
+        return response.status(403).json({ error: "UnAuthorized" });
       }
       return document.delete();
     })
